+++ web/app/ib/IB.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import R from 'ramda'
-import {LukionOppiaineEditor} from '../lukio/Lukio'
+import {LukionOppiaineEditor} from '../lukio/LukionOppiaineEditor'
 import {LukionOppiaineetTableHead} from '../lukio/fragments/LukionOppiaineetTable'
 import {modelData, modelLookup} from '../editor/EditorModel'
 import {FootnoteDescriptions} from '../components/footnote'
